sets up Calendar, MonthlyCalendar and Day components to dynamically generate monthly calendar based on number of days in month
